and finally! pool is stored in the thread!
for this reason, we don't need to pass it to the resolvers anymore.

Each session is also responsible for passing the connections it wants to
close (pool assumes all connections if nothing is passed); however, bear
in mind that connections are shared; if stuff gets preempted in the same
thread (fibers?), it might needlessly close connections and reopen them
again; a better solution has to be further investigated for this.