moved total timeout setup from request to the connection
connections are there for the lifecycle of a session, and that's where
the user expects it to pick up. At the request level is a bit weird, as
the total timeout will be unique for every request, but not all requests
fire at the same time, whereas, when a user sets total time, he expects
the whole httpx block to execute in that time.