select call rearchitected
we now go with the timer interval if it's short than the connection
timeouts. this way, timers will fire earlier if necessary.
Also, if there are timers ready to fire, instead of select-then-fire,
which will bypass fire if there are timeout errors (very likely), we now
fire the timers and restart the select process.