connection: perform all io in #consume
when condensing all io in the same method, one can now infer whether the
socket can process read/write still, thereby not needlessly sending it
to IO.select. This makes it for more efficient usage of cpu time.

Now we just need to do the same with the connect phase...