fixed the endless loop on exception for good
type check errors were sometimes causing the request loop to stall. This
was due to it not being handled in pool, and the respective requests not
being reaped. On exit, the session would try to close these connections,
but they weren't available anymore, causing it to happen. We were
already doing this for Interrupt, we're just extending it for all
non-StandarError exceptions