maintaining own fork of ruby-tls
The ruby-tls gem hasn't received an update in 3 years, and haven't
actioned on issues that old, which leads me to believe it was abandoned.
Therefore, I now maintain this fork, and this allows me to quickly
identify issues.

The flow within the IO wrapper has been dealt with as well