fixing double registry for DoH resolver
Some debugging surfaced that the resolving step of doh contains 2
registered connections in the selector, instead of just one. That's
because both the resolver and its inherent connection are registered.

Fixed it by marking the resolver as closed, in the same way we do for
the system resolver, which also nnever gets registered in the selector.