http2 fix: force connection termination if handshake hasn't transitioned
closing pools was stalling due to still-to-be-closed HTTP/2 connections.
this seemed to happened because connections transitioned to closing
state before having opened the HTTP/2 connection, so handshake was
skipped. Given that #close calls at the end of the session, we now check
if the HTTP/2 connection is closed already AND there are close handshake
bytes to send, after which we force-terminate the connection from the
client side.

This might have introduced because we now introduce idle connections
into the pool, what wasn't happening before (i.e. being introduced only
after having an IP to connect to).