fix: allow dns queries with ipv6 link-local address
An issue was mis-identified for years, where it was thought that IPv6
addresses with the net interface suffix, attributed by the LAN, was
invalid; this was wrong, as that's an ip address with a zone identifier.
`ipaddr` has since fixed its support, which invalidated a
native resolver patch, and the error surfaced when building a URI
component for the UDP factory, because `uri` does not support zone
identifiers.

A patch was therefore added to the URI component, to allow setting up
the host while not validating it, as that was the shortest path to
fixing it right now.