do not rely on send being called just once
the sentry and datadog plugins have been wrongly relying on the
assumption that #send is called just once, when in fact, it can be
called multiple times, both for conn exhaustion, as well as conn merging
(coalescing + happy eyeballs) scenarios.

Because of this, their "on response" callback could be set multiple times, which was confusing. So this fixes the behaviour.

Fixes #228