fix: ssl handshake correct handling of ip addresses
besides not setting session sni hostname, which it was already doing,
the verify_hostname is set to false to avoid warnings, and the
post_connection_check is still allowed to proceed, to check that the
certificate returned includes the IP address.

port of the similar net-http change found
[here](https://github.com/ruby/net-http/commit/fa68e64beee8ea3a9419b822ca10202822f69b5e)

also ommitting certain steps in the initializer if the ssl socket is
initiated outside of the httpx context and passed as an option.