fix: on altsvc processing, defer termination to when no more requests
a connection may be in the middle of processing multiple requests; once
an alt-svc header was received, while routing of pending data was done,
the current state of parser wasn't, but the connection has been
terminated and unlinked, leaving it in a state where errors may happen
if relying on attachment to a selector.

this fixes it by deferring termination to the point where the loop is
done with the current state of the connection; it does so by storing a
link to the altsvc connection.