fix potential session dereferencing in connection
there were a couple of situations where a connection would be
disconnected from the current session, and immediately picked up by a
separate thread, before the block of execution would reassign it again
to the current session, causing situations where the connection selector
didn't match the selector it was in. This gets exacerbated in persistent
connections.

this fixes it by relying in the previous simplification of the
connection lifecycle from the last commits; before resetting the
connection, parser pending requests get sent back to the connection; if
there are pending requests, then the connection won't be
closed/disconnected before being forced to the idle state, which ensures
it doesn't leave the current selector; if no pending requests, then it
disconnects, and must be picked up from the queue in order to be used
again.