fix: close only selected connections when an error occurs during tick
this removes the exception handlers on the loop, which was buggy;
besides force-closing connections which shouldn't be closed, it'd also
do it in an unsafe, where the selector selectables would be traversed
and force closed, which mutated the selectables listed being updated
(this could leave selectables hanging there); this becomes more
problematic for persistent connections.

this was fixed by rescuing at the wait/select call level, and traversing
the list of IOs being inspected.

another fix was propagation to resolvers which hold the connections,
which was not happening in the case of forceful termination of
resolvers, which didn't implement #force_reset. This was fixed by moving
to #force_close and implementing it for resolvers, which drops
connections and propagates #force_close.