stream: build_request(should mark a request as stream
from now on (and this is a breaking change!), #build_request should have
a `stream: true` parameter when the request is supposed to be a stream.

a warning is emitted if users are still using it the old fashioned way.

The reason is because the options object for the stream request MUST be
marked with the stream option, and there is no way to enforce that with
the previous API.

Closes https://github.com/HoneyryderChuck/httpx/issues/113