fix: force-close an IO which got an IOError while waiting for events
this error was being handled as in all other cases, and when the
connection was HTTP/2, the termination handshake bytes were buffered,
thereby entering a rabbit hole where it had bytes to flush, a closed
socket, and in the closing state.

the solution is to allow the error to bubble to requests and force-close
the socket.

Closes https://github.com/HoneyryderChuck/httpx/issues/120