auth plugin: pass block to dynamically generate token
in use cases where a token needs to be renewed, this plugin is useless,
as it's just a proxy to generate a new session, and the user has to deal
with the boilerplate.

This way, users can pass a block to the function, which will be
evaluated when building the request; the block can therefore contain
token refreshing logic.